fix(bevy)!: compose overlays after post-processing
Render Dear ImGui into Bevy's current unsampled target with a one-sample pipeline and expose stable before/overlay/after ordering phases. Add optional Bevy UI ordering, real GPU readback coverage, and split render-world responsibilities into focused modules.

BREAKING CHANGE: ImguiPipelineKey no longer exposes sample_count; Dear ImGui overlays always target the current single-sample attachment.
